+++ prtpy/partitioning/alternatives/utils.py
